Add iOS keyboard auto show on reply
More cleanup
