+++ src/App.tsx
@@ -17,12 +17,7 @@ import "@ionic/react/css/flex-utils.css";
 import "@ionic/react/css/display.css";
 
 import { StoreProvider } from "./store";
-import {
-  getAndroidNavMode,
-  isAndroid,
-  isInstalled,
-  isNative,
-} from "./helpers/device";
+import { getAndroidNavMode, isInstalled } from "./helpers/device";
 import TabbedRoutes from "./TabbedRoutes";
 import Auth from "./Auth";
 import { AppContextProvider } from "./features/auth/AppContext";
@@ -32,18 +27,15 @@ import { UpdateContextProvider } from "./pages/settings/update/UpdateContext";
 import GlobalStyles from "./GlobalStyles";
 import ConfigProvider from "./services/app";
 import { getDeviceMode } from "./features/settings/settingsSlice";
-import { SafeArea, SafeAreaInsets } from "capacitor-plugin-safe-area";
-import { StatusBar } from "@capacitor/status-bar";
-import { Keyboard } from "@capacitor/keyboard";
 import { TabContextProvider } from "./TabContext";
 import { NavModes } from "capacitor-android-nav-mode";
 import { TextRecoveryStartupPrompt } from "./helpers/useTextRecovery";
 import { ErrorBoundary } from "react-error-boundary";
 import AppCrash from "./AppCrash";
 
-import "./statusTap";
+import "./listeners";
 
-// index.tsx ensurxes android nav mode resolves before app is rendered
+// index.tsx ensures android nav mode resolves before app is rendered
 (async () => {
   let navMode;
   try {
@@ -55,41 +47,13 @@ import "./statusTap";
   setupIonicReact({
     rippleEffect: false,
     mode: getDeviceMode(),
-    statusTap: false, // custom implementation statusTap.ts
+    statusTap: false, // custom implementation listeners/statusTap.ts
     swipeBackEnabled:
       isInstalled() &&
       getDeviceMode() === "ios" &&
       navMode !== NavModes.Gesture,
   });
 })();
-
-// Android safe area inset management is bad, we have to do it manually
-if (isNative() && isAndroid()) {
-  let keyboardShowing = false;
-
-  const updateInsets = ({ insets }: SafeAreaInsets) => {
-    for (const [key, value] of Object.entries(insets)) {
-      document.documentElement.style.setProperty(
-        `--ion-safe-area-${key}`,
-        // if keyboard open, assume no safe area inset
-        `${keyboardShowing && key === "bottom" ? 0 : value}px`,
-      );
-    }
-  };
-
-  SafeArea.getSafeAreaInsets().then(updateInsets);
-  SafeArea.addListener("safeAreaChanged", updateInsets);
-  StatusBar.setOverlaysWebView({ overlay: true });
-
-  Keyboard.addListener("keyboardWillShow", () => {
-    keyboardShowing = true;
-    SafeArea.getSafeAreaInsets().then(updateInsets);
-  });
-  Keyboard.addListener("keyboardWillHide", () => {
-    keyboardShowing = false;
-    SafeArea.getSafeAreaInsets().then(updateInsets);
-  });
-}
 
 export default function App() {
   return (
