+++ src/listeners/index.ts
@@ -1,0 +1,3 @@
+import "./androidSafeArea";
+import "./keyboardPageResizer";
+import "./statusTap";
